Platformer level - layout and text object trigger updates
I added more room to the Platformer level image so there is an area to make holes.
I updated the text object trigger for "Group play" so it triggers earlier, and then I duplicated it so we can use the "Make holes and push them in" line again later in the level.